Changed way we encode VASM.
- VASM encoded as three address code to make conversions easier.
- This feature is not finished yet, committing so that I may work on other computer.
- Also takes up less memory now (no memory wasted on whitespace ascii chars).

+++ Parser/parser.h
@@ -37,7 +37,7 @@ TreeNode
 	struct TreeNode *right;
 	bool reg;
 	size_t rule_number;
-	int register_number;
+	size_t register_number;
 }
 TreeNode;
